a draft refactoring of the modules
